services/kepegawaian/requirements.txt
Flask==2.0.1
mysql-connector-python==8.0.26
prometheus-flask-exporter==0.23.2
Werkzeug==2.0.3